move newsletter components to lib/newsletter
Note: Components in lib/components should be free of app dependencies
so that can later be extracted to a separate repo. The newsletter
components depend on app-specific routes / form handlers / API client.
