+++ src/lib/components/index.ts
@@ -22,14 +22,12 @@ export { default as Logo } from './Logo.svelte';
 export { default as Menu } from './Menu.svelte';
 export { default as MenuItem } from './MenuItem.svelte';
 export { default as NavPanel } from './NavPanel.svelte';
-export { default as NewsletterOptInBanner } from './NewsletterOptInBanner.svelte';
 export { default as PageHeader } from './PageHeader.svelte';
 export { default as PageHeading } from './PageHeading.svelte';
 export { default as Section } from './Section.svelte';
 export { default as SmartContractWidget } from './SmartContractWidget.svelte';
 export { default as SocialIcon } from './SocialIcon.svelte';
 export { default as SourceCode } from './SourceCode.svelte';
-export { default as SubscribeForm } from './SubscribeForm.svelte';
 export { default as SummaryBox } from './SummaryBox.svelte';
 export { default as Tabs } from './Tabs.svelte';
 export { default as TextInput } from './TextInput.svelte';
